+++ CHANGES.md
@@ -1,6 +1,17 @@
 twilio-php Changelog
 ====================
 
+[2021-09-22] Version 6.28.3
+---------------------------
+**Events**
+- Add segment sink
+
+**Messaging**
+- Add post_approval_required attribute in GET us_app_to_person_usecase api response
+- Add Identity Status, Russell 3000, Tax Exempt Status and Should Skip SecVet fields for Brand Registrations
+- Add Should Skip Secondary Vetting optional flag parameter to create Brand API
+
+
 [2021-09-08] Version 6.28.2
 ---------------------------
 **Api**